VIDEO: fix seeking to first/second frame if it exceeds uint16

+++ video/flic_decoder.h
@@ -91,8 +91,8 @@ protected:
 		int _curFrame;
 		bool _atRingFrame;
 
-		uint16 _offsetFrame1;
-		uint16 _offsetFrame2;
+		uint32 _offsetFrame1;
+		uint32 _offsetFrame2;
 		byte *_palette;
 		mutable bool _dirtyPalette;
